@@ -6,14 +6,11 @@
 		"dev": "vite dev",
 		"build": "vite build",
 		"preview": "vite preview",
-		"test": "pnpm run test:unit && pnpm run test:integration",
 		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
-		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
-		"lint": "prettier --plugin-search-dir . --check . && eslint .",
-		"format": "prettier --plugin-search-dir . --write .",
-		"test:integration": "playwright test",
+		"lint": "prettier . --check . && eslint .",
+		"format": "prettier . --write .",
 		"test:unit": "vitest run",
-		"gh-pages": "pnpm run build && npx gh-pages -d build",
+		"test:integration": "playwright test",
 		"prepare": "husky"
 	},
 	"devDependencies": {
@@ -21,32 +18,32 @@
 		"@eslint/eslintrc": "^3.3.1",
 		"@eslint/js": "^9.33.0",
 		"@eslint/json": "^0.13.1",
-		"@playwright/test": "1.55.0",
-		"@sveltejs/adapter-auto": "^3.1.1",
-		"@sveltejs/adapter-static": "^3.0.1",
-		"@sveltejs/kit": "^2.34.1",
-		"@sveltejs/vite-plugin-svelte": "^3.1.2",
-		"@testing-library/svelte": "^4.0.3",
+		"@playwright/test": "^1.55.0",
+		"@sveltejs/adapter-auto": "^6.1.0",
+		"@sveltejs/adapter-static": "^3.0.9",
+		"@sveltejs/kit": "^2.36.1",
+		"@sveltejs/vite-plugin-svelte": "^6.1.3",
+		"@testing-library/svelte": "^5.2.8",
 		"@typescript-eslint/eslint-plugin": "^8.40.0",
 		"@typescript-eslint/parser": "^8.40.0",
 		"eslint": "^9.33.0",
 		"eslint-config-prettier": "^10.1.8",
 		"eslint-plugin-svelte": "^3.11.0",
-		"gh-pages": "^5.0.0",
 		"globals": "^16.3.0",
 		"husky": "^9.1.7",
 		"jiti": "^2.5.1",
-		"jsdom": "^22.1.0",
-		"prettier": "^2.8.0",
-		"prettier-plugin-svelte": "^2.10.1",
-		"svelte": "^4.2.20",
-		"svelte-check": "^4.1.0",
-		"tslib": "^2.4.1",
-		"typescript": "^5.0.0",
+		"jsdom": "^26.1.0",
+		"prettier": "^3.6.2",
+		"prettier-plugin-svelte": "^3.4.0",
+		"svelte": "^5.38.2",
+		"svelte-check": "^4.3.1",
+		"svelte-eslint-parser": "^1.3.1",
+		"tslib": "^2.8.1",
+		"typescript": "^5.9.2",
 		"typescript-eslint": "^8.40.0",
-		"vite": "^5.4.19",
+		"vite": "^7.1.3",
 		"vite-plugin-pwa": "^1.0.3",
-		"vitest": "^0.33.0"
+		"vitest": "^3.2.4"
 	},
 	"type": "module",
 	"dependencies": {
