@@ -1,10 +1,10 @@
 <script lang="ts">
-    import {createEventDispatcher} from 'svelte';
-    import Dialog from '$lib/components/pure/dialog/Dialog.svelte';
-    import Heading3 from '$lib/components/pure/heading/Heading3.svelte';
-    import TextButton from '$lib/components/pure/text-button/TextButton.svelte';
+	import { createEventDispatcher } from 'svelte';
+	import Dialog from '$lib/components/pure/dialog/Dialog.svelte';
+	import Heading3 from '$lib/components/pure/heading/Heading3.svelte';
+	import TextButton from '$lib/components/pure/text-button/TextButton.svelte';
 
-    export let dialog: HTMLDialogElement;
+	export let dialog: HTMLDialogElement;
 	export let title: string;
 	export let warning: string;
 	export let labelConfirm: string;
