@@ -1,6 +1,6 @@
 <script>
-    import IconLink from '$lib/components/pure/icon-link/IconLink.svelte';
-    import DiscordIcon from './DiscordIcon.svelte';
+	import IconLink from '$lib/components/pure/icon-link/IconLink.svelte';
+	import DiscordIcon from './DiscordIcon.svelte';
 </script>
 
 <IconLink title="Discord Server" href="https://discord.gg/yyJFG5PVBZ" newtab>
