@@ -1,11 +1,11 @@
 <script lang="ts">
-    import {formatVideoTime} from '$lib/logic/format-time';
-    import {isSpace} from '$lib/logic/keyboard-utils';
-    import {onDestroy, onMount} from 'svelte';
-    import {browser} from '$app/environment';
-    import PlayButton from '../button-play/PlayButton.svelte';
+	import { formatVideoTime } from '$lib/logic/format-time';
+	import { isSpace } from '$lib/logic/keyboard-utils';
+	import { onDestroy, onMount } from 'svelte';
+	import { browser } from '$app/environment';
+	import PlayButton from '../button-play/PlayButton.svelte';
 
-    const SKIP_TIME = 5;
+	const SKIP_TIME = 5;
 
 	export let src: string;
 	export let poster: string;
@@ -131,7 +131,7 @@
 			on:dblclick|stopPropagation|preventDefault={skip}
 			preload="metadata"
 			style={`aspect-ratio: ${width} / ${height}`}
-		></video>
+		/>
 		<span class:hide={intentHideOverlay} class="hidable">{formatVideoTime(timeLeft)}</span>
 		<input
 			bind:value={currentTime}
