@@ -1,11 +1,11 @@
 <script lang="ts">
-    import {getComments} from '$lib/logic/api-client/ApiClient';
-    import LoadingAnimation from '$lib/components/pure/loading-animation/LoadingAnimation.svelte';
-    import Comment from '$lib/components/kurosearch/post-comment/Comment.svelte';
-    import apiKey from '$lib/store/api-key-store';
-    import userId from '$lib/store/user-id-store';
+	import { getComments } from '$lib/logic/api-client/ApiClient';
+	import LoadingAnimation from '$lib/components/pure/loading-animation/LoadingAnimation.svelte';
+	import Comment from '$lib/components/kurosearch/post-comment/Comment.svelte';
+	import apiKey from '$lib/store/api-key-store';
+	import userId from '$lib/store/user-id-store';
 
-    export let post: kurosearch.Post;
+	export let post: kurosearch.Post;
 </script>
 
 {#await getComments(post.id, $apiKey, $userId)}
