@@ -1,7 +1,7 @@
 <script lang="ts">
-    import {formatCreatedAt} from '$lib/logic/format-relative-time';
+	import { formatCreatedAt } from '$lib/logic/format-relative-time';
 
-    export let value: number | string;
+	export let value: number | string;
 </script>
 
 <span>{formatCreatedAt(value)}</span>
