@@ -1,12 +1,12 @@
 <script lang="ts">
-    import {getPostId} from '$lib/logic/id-utils';
-    import resultColumns from '$lib/store/result-columns-store';
-    import results from '$lib/store/results-store';
-    import FullscreenPost from '../fullscreen-post/FullscreenPost.svelte';
-    import MosaicPost from '../post/MosaicPost.svelte';
-    import SingleColumnPost from '../post/SingleColumnPost.svelte';
+	import { getPostId } from '$lib/logic/id-utils';
+	import resultColumns from '$lib/store/result-columns-store';
+	import results from '$lib/store/results-store';
+	import FullscreenPost from '../fullscreen-post/FullscreenPost.svelte';
+	import MosaicPost from '../post/MosaicPost.svelte';
+	import SingleColumnPost from '../post/SingleColumnPost.svelte';
 
-    let fullscreenIndex: undefined | number;
+	let fullscreenIndex: undefined | number;
 
 	const exitFullscreen = (event: CustomEvent<number>) => {
 		const post = $results.posts[event.detail];
