@@ -1,7 +1,7 @@
 <script lang="ts">
-    import {formatCount} from '$lib/logic/format-count';
+	import { formatCount } from '$lib/logic/format-count';
 
-    export let value: number;
+	export let value: number;
 </script>
 
 <span data-testid="score" class="codicon codicon-heart">{formatCount(value)}</span>
