@@ -1,9 +1,9 @@
 <script lang="ts">
-    import {formatCount} from '$lib/logic/format-count';
-    import {formatTagname} from '$lib/logic/format-tag';
-    import {createEventDispatcher} from 'svelte';
+	import { formatCount } from '$lib/logic/format-count';
+	import { formatTagname } from '$lib/logic/format-tag';
+	import { createEventDispatcher } from 'svelte';
 
-    const dispatch = createEventDispatcher();
+	const dispatch = createEventDispatcher();
 
 	export let suggestion: kurosearch.Suggestion;
 	export let selected = false;
@@ -11,20 +11,20 @@
 
 <!-- svelte-ignore a11y-no-noninteractive-element-to-interactive-role -->
 <!-- svelte-ignore a11y-click-events-have-key-events -->
-<li
+<button
 	role="button"
 	tabindex="0"
 	on:click={() => dispatch('click', suggestion)}
 	title={suggestion.label}
 	class:selected
 >
-	<i class={suggestion.type === 'supertag' ? 'codicon codicon-star-full' : 'codicon codicon-tag'} />
+	<i class={suggestion.type === 'supertag' ? 'codicon codicon-star-full' : 'codicon codicon-tag'} ></i>
 	<span class="tag-name">{formatTagname(suggestion.label)}</span>
 	<span class="tag-count">{formatCount(suggestion.count)}</span>
-</li>
+</button>
 
 <style>
-	li {
+	button {
 		display: grid;
 		align-items: center;
 		height: var(--line-height-small);
@@ -35,7 +35,7 @@
 		user-select: none;
 	}
 
-	li:focus,
+	button:focus,
 	.selected {
 		background-color: var(--background-2);
 		outline: none;
@@ -48,11 +48,11 @@
 	}
 
 	@media (hover: hover) {
-		li {
+		button {
 			transition: background-color var(--default-transition-behaviour);
 		}
 
-		li:hover {
+		button:hover {
 			background-color: var(--background-2);
 		}
 	}
