@@ -1,7 +1,7 @@
 <script lang="ts">
-    import IconLink from './IconLink.svelte';
+	import IconLink from './IconLink.svelte';
 
-    export let title: string;
+	export let title: string;
 	export let href: string;
 	export let icon: string;
 	export let newtab = false;
