@@ -32,7 +32,7 @@ export const BLOCKING_GROUP_TAGS: Record<kurosearch.BlockingGroup, readonly stri
 			'scared',
 			'forced'
 		]),
-		Gore: Object.freeze(['gore', 'necrophilia', 'amputee', 'guro', 'blood']),
+		Gore: Object.freeze(['gore', 'necrophilia', 'amputee', 'guro', 'blood', 'amputed*']),
 		Scat: Object.freeze(['scat', 'diaper', 'fart']),
 		Vore: Object.freeze(['vore'])
 	});
