@@ -13,7 +13,7 @@ import { createSearchableTag, createSupertag } from '../tag-utils';
 import { logFirestoreRead, logFirestoreWrite } from './analytics';
 
 export const getSettingsAndSupertags = async () => {
-	logFirestoreRead();
+	await logFirestoreRead();
 	const settings = await getPreferences();
 	const supertags = await getSupertags();
 
@@ -102,7 +102,7 @@ const setSupertags = async (supertags: kurosearch.Supertag[]) => {
 };
 
 export const saveSettingsAndSupertags = async (settings: any, supertags: kurosearch.Supertag[]) => {
-	logFirestoreWrite();
+	await logFirestoreWrite();
 	await setPreferences(settings);
 	await setSupertags(supertags);
 };
