@@ -1,29 +1,29 @@
 <script lang="ts">
-    import {browser} from '$app/environment';
-    import SearchError from '$lib/components/kurosearch/error-search/SearchError.svelte';
-    import NoMoreResults from '$lib/components/kurosearch/results/NoMoreResults.svelte';
-    import ResultHeader from '$lib/components/kurosearch/results/ResultHeader.svelte';
-    import Results from '$lib/components/kurosearch/results/Results.svelte';
-    import ZeroResults from '$lib/components/kurosearch/results/ZeroResults.svelte';
-    import ScrollUpButton from '$lib/components/pure/button-scroll-up/ScrollUpButton.svelte';
-    import IntersectionDetector from '$lib/components/pure/intersection-detector/IntersectionDetector.svelte';
-    import LoadingAnimation from '$lib/components/pure/loading-animation/LoadingAnimation.svelte';
-    import TextButton from '$lib/components/pure/text-button/TextButton.svelte';
-    import {logSearch} from '$lib/logic/firebase/analytics';
-    import {SearchBuilder} from '$lib/logic/search-builder';
-    import activeSupertags from '$lib/store/active-supertags-store';
-    import activeTags from '$lib/store/active-tags-store';
-    import blockedContent from '$lib/store/blocked-content-store';
-    import filter from '$lib/store/filter-store';
-    import resultColumns from '$lib/store/result-columns-store';
-    import results from '$lib/store/results-store';
-    import sort from '$lib/store/sort-store';
-    import {onDestroy, onMount} from 'svelte';
-    import SearchForm from './SearchForm.svelte';
-    import apiKey from '$lib/store/api-key-store';
-    import userId from '$lib/store/user-id-store';
+	import { browser } from '$app/environment';
+	import SearchError from '$lib/components/kurosearch/error-search/SearchError.svelte';
+	import NoMoreResults from '$lib/components/kurosearch/results/NoMoreResults.svelte';
+	import ResultHeader from '$lib/components/kurosearch/results/ResultHeader.svelte';
+	import Results from '$lib/components/kurosearch/results/Results.svelte';
+	import ZeroResults from '$lib/components/kurosearch/results/ZeroResults.svelte';
+	import ScrollUpButton from '$lib/components/pure/button-scroll-up/ScrollUpButton.svelte';
+	import IntersectionDetector from '$lib/components/pure/intersection-detector/IntersectionDetector.svelte';
+	import LoadingAnimation from '$lib/components/pure/loading-animation/LoadingAnimation.svelte';
+	import TextButton from '$lib/components/pure/text-button/TextButton.svelte';
+	import { logSearch } from '$lib/logic/firebase/analytics';
+	import { SearchBuilder } from '$lib/logic/search-builder';
+	import activeSupertags from '$lib/store/active-supertags-store';
+	import activeTags from '$lib/store/active-tags-store';
+	import blockedContent from '$lib/store/blocked-content-store';
+	import filter from '$lib/store/filter-store';
+	import resultColumns from '$lib/store/result-columns-store';
+	import results from '$lib/store/results-store';
+	import sort from '$lib/store/sort-store';
+	import { onDestroy, onMount } from 'svelte';
+	import SearchForm from './SearchForm.svelte';
+	import apiKey from '$lib/store/api-key-store';
+	import userId from '$lib/store/user-id-store';
 
-    let loading = false;
+	let loading = false;
 	let error: Error | undefined;
 	let nextFocus = 0;
 
@@ -65,14 +65,14 @@
 		results.reset();
 		nextFocus = 0;
 
-		executeSearch(async () => {
+		await executeSearch(async () => {
 			const [page, count] = await createDefaultSearch().getPageAndCount();
 			results.addPage(page, count);
 		});
 	};
 
 	const getNextPage = async () => {
-		executeSearch(async () => {
+		await executeSearch(async () => {
 			const page = await createDefaultSearch().getPage();
 			results.addPage(page);
 		});
