@@ -1,7 +1,7 @@
 <script lang="ts">
-    import {base} from '$app/paths';
+	import { base } from '$app/paths';
 
-    let details = false;
+	let details = false;
 </script>
 
 <div>
