@@ -1,24 +1,24 @@
 <script lang="ts">
-    import {browser} from '$app/environment';
-    import KurosearchTitle from '$lib/components/kurosearch/kurosearch-title/KurosearchTitle.svelte';
-    import Searchbar from '$lib/components/kurosearch/searchbar/Searchbar.svelte';
-    import ActiveTagList from '$lib/components/kurosearch/tag-list/ActiveTagList.svelte';
-    import LoadingAnimation from '$lib/components/pure/loading-animation/LoadingAnimation.svelte';
-    import TextButton from '$lib/components/pure/text-button/TextButton.svelte';
-    import {getTagSuggestions} from '$lib/logic/api-client/ApiClient';
-    import {getTagDetails} from '$lib/logic/api-client/tags/tags';
-    import {nextModifier} from '$lib/logic/modifier-utils';
-    import {addHistory} from '$lib/logic/use/onpopstate';
-    import activeSupertags from '$lib/store/active-supertags-store';
-    import activeTags from '$lib/store/active-tags-store';
-    import results from '$lib/store/results-store';
-    import supertags from '$lib/store/supertags-store';
-    import {createEventDispatcher, onDestroy, onMount} from 'svelte';
-    import CreateSupertagDialog from '$lib/components/kurosearch/dialog-create-supertag/CreateSupertagDialog.svelte';
-    import apiKey from '$lib/store/api-key-store';
-    import userId from '$lib/store/user-id-store';
+	import { browser } from '$app/environment';
+	import KurosearchTitle from '$lib/components/kurosearch/kurosearch-title/KurosearchTitle.svelte';
+	import Searchbar from '$lib/components/kurosearch/searchbar/Searchbar.svelte';
+	import ActiveTagList from '$lib/components/kurosearch/tag-list/ActiveTagList.svelte';
+	import LoadingAnimation from '$lib/components/pure/loading-animation/LoadingAnimation.svelte';
+	import TextButton from '$lib/components/pure/text-button/TextButton.svelte';
+	import { getTagSuggestions } from '$lib/logic/api-client/ApiClient';
+	import { getTagDetails } from '$lib/logic/api-client/tags/tags';
+	import { nextModifier } from '$lib/logic/modifier-utils';
+	import { addHistory } from '$lib/logic/use/onpopstate';
+	import activeSupertags from '$lib/store/active-supertags-store';
+	import activeTags from '$lib/store/active-tags-store';
+	import results from '$lib/store/results-store';
+	import supertags from '$lib/store/supertags-store';
+	import { createEventDispatcher, onDestroy, onMount } from 'svelte';
+	import CreateSupertagDialog from '$lib/components/kurosearch/dialog-create-supertag/CreateSupertagDialog.svelte';
+	import apiKey from '$lib/store/api-key-store';
+	import userId from '$lib/store/user-id-store';
 
-    const dispatch = createEventDispatcher();
+	const dispatch = createEventDispatcher();
 	const submit = () => dispatch('submit');
 
 	export let loading: boolean;
