@@ -1,12 +1,12 @@
 <script>
-    import DetailedTag from '$lib/components/kurosearch/tag-detailed/DetailedTag.svelte';
-    import SimpleTag from '$lib/components/kurosearch/tag-simple/SimpleTag.svelte';
-    import ScrollUpButton from '$lib/components/pure/button-scroll-up/ScrollUpButton.svelte';
-    import Heading1 from '$lib/components/pure/heading/Heading1.svelte';
-    import Heading3 from '$lib/components/pure/heading/Heading3.svelte';
-    import SummaryCard from '$lib/components/pure/summary-card/SummaryCard.svelte';
-    import {MODIFIERS_ICONS} from '$lib/logic/tag-modifier-data';
-    import {createModifiedTag, createTag} from '$lib/logic/tag-utils';
+	import DetailedTag from '$lib/components/kurosearch/tag-detailed/DetailedTag.svelte';
+	import SimpleTag from '$lib/components/kurosearch/tag-simple/SimpleTag.svelte';
+	import ScrollUpButton from '$lib/components/pure/button-scroll-up/ScrollUpButton.svelte';
+	import Heading1 from '$lib/components/pure/heading/Heading1.svelte';
+	import Heading3 from '$lib/components/pure/heading/Heading3.svelte';
+	import SummaryCard from '$lib/components/pure/summary-card/SummaryCard.svelte';
+	import { MODIFIERS_ICONS } from '$lib/logic/tag-modifier-data';
+	import { createModifiedTag, createTag } from '$lib/logic/tag-utils';
 </script>
 
 <svelte:head>
