@@ -1,21 +1,21 @@
 <script>
-    import {browser} from '$app/environment';
-    import Gif from '$lib/components/kurosearch/media-gif/Gif.svelte';
-    import Video from '$lib/components/kurosearch/media-video/Video.svelte';
-    import Comments from '$lib/components/kurosearch/post-comment/Comments.svelte';
-    import Rating from '$lib/components/kurosearch/rating/Rating.svelte';
-    import RelativeTime from '$lib/components/kurosearch/relative-time/RelativeTime.svelte';
-    import Score from '$lib/components/kurosearch/score/Score.svelte';
-    import ExternalSource from '$lib/components/kurosearch/source-external/ExternalSource.svelte';
-    import Rule34Source from '$lib/components/kurosearch/source-rule34/Rule34Source.svelte';
-    import PostDetailsTagList from '$lib/components/kurosearch/tag-list/PostDetailsTagList.svelte';
-    import LoadingAnimation from '$lib/components/pure/loading-animation/LoadingAnimation.svelte';
-    import PreviewedImage from '$lib/components/pure/smart-image/PreviewedImage.svelte';
-    import {getPost} from '$lib/logic/api-client/posts/posts';
-    import {getVideoSources, isLoop} from '$lib/logic/media-utils';
-    import alwaysLoop from '$lib/store/always-loop-store';
+	import { browser } from '$app/environment';
+	import Gif from '$lib/components/kurosearch/media-gif/Gif.svelte';
+	import Video from '$lib/components/kurosearch/media-video/Video.svelte';
+	import Comments from '$lib/components/kurosearch/post-comment/Comments.svelte';
+	import Rating from '$lib/components/kurosearch/rating/Rating.svelte';
+	import RelativeTime from '$lib/components/kurosearch/relative-time/RelativeTime.svelte';
+	import Score from '$lib/components/kurosearch/score/Score.svelte';
+	import ExternalSource from '$lib/components/kurosearch/source-external/ExternalSource.svelte';
+	import Rule34Source from '$lib/components/kurosearch/source-rule34/Rule34Source.svelte';
+	import PostDetailsTagList from '$lib/components/kurosearch/tag-list/PostDetailsTagList.svelte';
+	import LoadingAnimation from '$lib/components/pure/loading-animation/LoadingAnimation.svelte';
+	import PreviewedImage from '$lib/components/pure/smart-image/PreviewedImage.svelte';
+	import { getPost } from '$lib/logic/api-client/posts/posts';
+	import { getVideoSources, isLoop } from '$lib/logic/media-utils';
+	import alwaysLoop from '$lib/store/always-loop-store';
 
-    const url = browser ? new URL(location.href) : undefined;
+	const url = browser ? new URL(location.href) : undefined;
 	const id = url && url.searchParams.has('id') ? Number(url.searchParams.get('id')) : undefined;
 	const src = url && url.searchParams.get('src');
 
