@@ -1,10 +1,10 @@
 <script lang="ts">
-    import {browser} from '$app/environment';
-    import Searchbar from '$lib/components/kurosearch/searchbar/Searchbar.svelte';
-    import LoadingAnimation from '$lib/components/pure/loading-animation/LoadingAnimation.svelte';
-    import {getTagSuggestions} from '$lib/logic/api-client/ApiClient';
+	import { browser } from '$app/environment';
+	import Searchbar from '$lib/components/kurosearch/searchbar/Searchbar.svelte';
+	import LoadingAnimation from '$lib/components/pure/loading-animation/LoadingAnimation.svelte';
+	import { getTagSuggestions } from '$lib/logic/api-client/ApiClient';
 
-    const name = browser ? new URL(document.location.href).searchParams.get('name') : undefined;
+	const name = browser ? new URL(document.location.href).searchParams.get('name') : undefined;
 
 	const loadTag = async (name: string) => {
 		const response = await fetch(
